rename navigation to toolmanager
