+++ examples/user_interfaces/toolmanager.py
@@ -8,8 +8,9 @@ class allows to:
 
 from __future__ import print_function
 import matplotlib
-matplotlib.use('GTK3Cairo')
-matplotlib.rcParams['toolbar'] = 'navigation'
+# matplotlib.use('GTK3Cairo')
+matplotlib.use('Tkagg')
+matplotlib.rcParams['toolbar'] = 'toolmanager'
 import matplotlib.pyplot as plt
 from matplotlib.backend_tools import ToolBase
 from gi.repository import Gtk, Gdk
@@ -27,11 +28,11 @@ class ListTools(ToolBase):
                                          'Tool description',
                                          'Keymap'))
         print('-' * 80)
-        tools = self.navigation.tools
+        tools = self.toolmanager.tools
         for name in sorted(tools.keys()):
             if not tools[name].description:
                 continue
-            keys = ', '.join(sorted(self.navigation.get_tool_keymap(name)))
+            keys = ', '.join(sorted(self.toolmanager.get_tool_keymap(name)))
             print("{0:12} {1:45} {2}".format(name,
                                              tools[name].description,
                                              keys))
@@ -39,7 +40,7 @@ class ListTools(ToolBase):
         print("Active Toggle tools")
         print("{0:12} {1:45}".format("Group", "Active"))
         print('-' * 80)
-        for group, active in self.navigation.active_toggle.items():
+        for group, active in self.toolmanager.active_toggle.items():
             print("{0:12} {1:45}".format(group, active))
 
 
@@ -61,15 +62,15 @@ fig = plt.figure()
 plt.plot([1, 2, 3])
 
 # Add the custom tools that we created
-fig.canvas.manager.navigation.add_tool('List', ListTools)
-fig.canvas.manager.navigation.add_tool('copy', CopyToolGTK3)
+fig.canvas.manager.toolmanager.add_tool('List', ListTools)
+# fig.canvas.manager.toolmanager.add_tool('copy', CopyToolGTK3)
 
 # Add an existing tool to new group `foo`.
 # It can be added as many times as we want
 fig.canvas.manager.toolbar.add_tool('zoom', 'foo')
 
 # Remove the forward button
-fig.canvas.manager.navigation.remove_tool('forward')
+fig.canvas.manager.toolmanager.remove_tool('forward')
 
 # To add a custom tool to the toolbar at specific location
 fig.canvas.manager.toolbar.add_tool('List', 'navigation', 1)
