+++ lib/matplotlib/rcsetup.py
@@ -168,7 +168,7 @@ validate_qt5 = ValidateInStrings('backend.qt5', ['PyQt5'])
 def validate_toolbar(s):
     validator = ValidateInStrings(
                 'toolbar',
-                ['None', 'toolbar2', 'navigation'],
+                ['None', 'toolbar2', 'toolmanager'],
                 ignorecase=True)
     return validator(s)
